pyproject.toml
[tool.poetry]
name = "privategpt"
version = "0.1.0"
description = ""
authors = ["Ivan Martinez <ivanmartit@gmail.com>"]
license = "Apache Version 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.274"
gpt4all = "^1.0.8"
chromadb = "^0.4.7"
llama-cpp-python = "0.1.78"
urllib3 = "^2.0.4"
PyMuPDF = "^1.23.1"
python-dotenv = "^1.0.0"
unstructured = "^0.10.8"
extract-msg = "^0.45.0"
tabulate = "^0.9.0"
pandoc = "^2.3"
pypandoc = "^1.11"
tqdm = "^4.66.1"
torch = ">=2.0.0, !=2.0.1"
sentence-transformers = "^2.2.2"
numba = "^0.57.0"
whisper-cpp-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
